refactor: 코드 리뷰 반영
- 구현체 1개인 객체의 인터페이스 제거
- 불필요한 코드 제거 및 수정

+++ src/main/java/org/c4marathon/assignment/bankaccount/dto/response/SavingAccountResponseDto.java
@@ -5,8 +5,8 @@ import lombok.Builder;
 @Builder
 public record SavingAccountResponseDto(
 	long accountPk,
-	int savingMoney,
-	int rate,
+	long savingMoney,
+	long rate,
 	String productName
 ) {
 }
